+++ tests/table/index.test.tsx
@@ -383,7 +383,7 @@ describe('BasicTable', () => {
     await waitForComponentToPaint(html, 1200);
 
     act(() => {
-      html.find('.ant-pro-table-toolbar-setting-item span.anticon-reload').simulate('click');
+      html.find('.ant-pro-table-list-toolbar-setting-item span.anticon-reload').simulate('click');
     });
 
     await waitForComponentToPaint(html, 1200);
@@ -417,7 +417,9 @@ describe('BasicTable', () => {
     await waitForComponentToPaint(html);
 
     act(() => {
-      html.find('.ant-pro-table-toolbar-setting-item span.anticon-column-height').simulate('click');
+      html
+        .find('.ant-pro-table-list-toolbar-setting-item span.anticon-column-height')
+        .simulate('click');
     });
     await waitForComponentToPaint(html);
     act(() => {
@@ -514,7 +516,9 @@ describe('BasicTable', () => {
     await waitForComponentToPaint(html, 1200);
 
     act(() => {
-      html.find('.ant-pro-table-toolbar-setting-item span.anticon-fullscreen').simulate('click');
+      html
+        .find('.ant-pro-table-list-toolbar-setting-item span.anticon-fullscreen')
+        .simulate('click');
     });
 
     await waitForComponentToPaint(html, 1200);
@@ -576,7 +580,9 @@ describe('BasicTable', () => {
     await waitForComponentToPaint(html, 600);
 
     act(() => {
-      html.find('.ant-pro-table-toolbar-setting-item span.anticon-fullscreen').simulate('click');
+      html
+        .find('.ant-pro-table-list-toolbar-setting-item span.anticon-fullscreen')
+        .simulate('click');
     });
     await waitForComponentToPaint(html, 1200);
 
@@ -584,7 +590,7 @@ describe('BasicTable', () => {
 
     act(() => {
       html
-        .find('.ant-pro-table-toolbar-setting-item span.anticon-fullscreen-exit')
+        .find('.ant-pro-table-list-toolbar-setting-item span.anticon-fullscreen-exit')
         .simulate('click');
     });
 
@@ -621,7 +627,9 @@ describe('BasicTable', () => {
     await waitForComponentToPaint(html, 1200);
 
     act(() => {
-      html.find('.ant-pro-table-toolbar-setting-item span.anticon-column-height').simulate('click');
+      html
+        .find('.ant-pro-table-list-toolbar-setting-item span.anticon-column-height')
+        .simulate('click');
     });
     await waitForComponentToPaint(html, 1200);
     act(() => {
@@ -697,7 +705,7 @@ describe('BasicTable', () => {
     await waitForComponentToPaint(html, 600);
 
     act(() => {
-      html.find('.ant-pro-table-toolbar-search input').simulate('change', {
+      html.find('.ant-pro-table-list-toolbar-search input').simulate('change', {
         target: {
           value: 'name',
         },
@@ -706,7 +714,7 @@ describe('BasicTable', () => {
 
     act(() => {
       html
-        .find('.ant-pro-table-toolbar-search input')
+        .find('.ant-pro-table-list-toolbar-search input')
         .simulate('keydown', { key: 'Enter', keyCode: 13 });
     });
 
@@ -740,7 +748,7 @@ describe('BasicTable', () => {
     await waitForComponentToPaint(html, 600);
 
     act(() => {
-      html.find('.ant-pro-table-toolbar-search input').simulate('change', {
+      html.find('.ant-pro-table-list-toolbar-search input').simulate('change', {
         target: {
           value: 'name',
         },
@@ -749,7 +757,7 @@ describe('BasicTable', () => {
 
     act(() => {
       html
-        .find('.ant-pro-table-toolbar-search input')
+        .find('.ant-pro-table-list-toolbar-search input')
         .simulate('keydown', { key: 'Enter', keyCode: 13 });
     });
