+++ tests/utils/index.test.tsx
@@ -3,12 +3,7 @@ import {
   parseValueToMoment,
   transformKeySubmitValue,
   isNil,
-  ListToolBar,
 } from '@ant-design/pro-utils';
-import { SettingOutlined, FullscreenOutlined } from '@ant-design/icons';
-import { Button } from 'antd';
-import * as React from 'react';
-import { mount } from 'enzyme';
 import moment, { Moment } from 'moment';
 
 describe('utils', () => {
@@ -228,138 +223,5 @@ describe('utils', () => {
     expect(isNil('')).toBe(false);
     expect(isNil({})).toBe(false);
     expect(isNil(true)).toBe(false);
-  });
-
-  it('ListToolBar onAction', () => {
-    const onAction = jest.fn();
-    const wrapper = mount(
-      <ListToolBar
-        actions={[
-          <Button>批量导入</Button>,
-          <Button
-            type="primary"
-            onClick={() => {
-              onAction('add');
-            }}
-          >
-            添加
-          </Button>,
-        ]}
-      />,
-    );
-    wrapper.find('button.ant-btn-primary').simulate('click');
-    expect(onAction).toHaveBeenLastCalledWith('add');
-  });
-
-  it('ListToolBar onSettingClick', () => {
-    const onClick = jest.fn();
-    const wrapper = mount(
-      <ListToolBar
-        settings={[
-          {
-            icon: <SettingOutlined />,
-            tooltip: '设置',
-            onClick,
-            key: 's-value',
-          },
-          {
-            icon: <FullscreenOutlined />,
-            tooltip: '全屏',
-          },
-        ]}
-      />,
-    );
-    wrapper.find('.anticon-setting').simulate('click');
-    expect(onClick).toHaveBeenLastCalledWith('s-value');
-    expect(wrapper.find('.ant-divider').length).toEqual(0);
-  });
-
-  it('ListToolBar search left', () => {
-    const onSearch = jest.fn();
-    const wrapper = mount(
-      <ListToolBar
-        search={{
-          placeholder: '自定义 placeholder',
-          onSearch,
-        }}
-      />,
-    );
-    const inputEle = wrapper.find('input');
-    inputEle.simulate('focus');
-    inputEle.simulate('change', { target: { value: 'input 值' } });
-    inputEle.simulate('keyDown', { keyCode: 13 });
-    expect(wrapper.find('.ant-pro-table-toolbar-left input').prop('value')).toEqual('input 值');
-    expect(onSearch).toHaveBeenCalled();
-    expect(wrapper.find('input').prop('placeholder')).toEqual('自定义 placeholder');
-  });
-
-  it('ListToolBar search right', () => {
-    const onSearch = jest.fn();
-    const wrapper = mount(
-      <ListToolBar
-        title="I am title"
-        search={{
-          placeholder: '自定义 placeholder',
-          onSearch,
-        }}
-      />,
-    );
-    const inputEle = wrapper.find('input');
-    inputEle.simulate('focus');
-    inputEle.simulate('change', { target: { value: 'input 值' } });
-    inputEle.simulate('keyDown', { keyCode: 13 });
-    expect(wrapper.find('.ant-pro-table-toolbar-right input').prop('value')).toEqual('input 值');
-    expect(onSearch).toHaveBeenCalled();
-    expect(wrapper.find('input').prop('placeholder')).toEqual('自定义 placeholder');
-  });
-
-  it('ListToolBar menu', () => {
-    const onChange = jest.fn();
-    const wrapper = mount(
-      <ListToolBar
-        menu={{
-          type: 'dropdown',
-          items: [
-            {
-              label: '全部事项',
-              key: 'all',
-            },
-            {
-              label: '已办事项',
-              key: 'done',
-            },
-          ],
-          onChange,
-        }}
-      />,
-    );
-    wrapper.find('.ant-pro-table-toolbar-dropdownmenu-label').simulate('click');
-    wrapper.find('.ant-dropdown-menu-item').at(1).simulate('click');
-
-    expect(onChange).toHaveBeenCalledWith('done', undefined);
-  });
-
-  it('ListToolBar inline menu', () => {
-    const onChange = jest.fn();
-    const wrapper = mount(
-      <ListToolBar
-        menu={{
-          type: 'inline',
-          items: [
-            {
-              label: '全部事项',
-              key: 'all',
-            },
-            {
-              label: '已办事项',
-              key: 'done',
-            },
-          ],
-          onChange,
-        }}
-      />,
-    );
-    wrapper.find('.ant-pro-table-toolbar-inlinemenu-item').at(1).simulate('click');
-    expect(onChange).toHaveBeenCalledWith('done', undefined);
   });
 });
